+++ core/api/service/chain/impl/chain_api_impl.cpp
@@ -64,7 +64,7 @@ namespace kagome::api {
     return results;
   }
 
-  outcome::result<bool> ChainApiImpl::unsubscribeNewHeads(uint32_t id) {
+  outcome::result<bool> ChainApiImpl::unsubscribeNewHeads(int64_t id) {
     if (auto api_service = api_service_.lock())
       return api_service->unsubscribeNewHeads(id);
